requirements.txt
boto3==1.12.26
Flask==1.1.1
opencv-python==4.2.0.32
Pillow==7.0.0
Werkzeug==0.16.1
